// src/sections/game/tile-match/game/stage-hud/types.ts
export interface MenuEditorSectionProps {
  editorActiveIndex?: number;
  editorLevels?: Array<{
    name: string;
    turnLimit?: number;
    grid: CellType[][];
    hint?: CellType[][][];
  }>;
  editorUpdateTurnLimit?: (turns: number) => void;
  editorMapType?: 'real' | 'test';
  setEditorMapType?: (type: 'real' | 'test') => void;
  changeMapType?: (type: 'real' | 'test') => void;
  playSound: (
    type: 'coin' | 'select' | 'start' | 'error' | 'match' | 'fall' | 'shoot' | 'break',
    muted: boolean
  ) => void;
  muted: boolean;
  togglePlayTest?: () => void;
  playTestMode: boolean;
  setIsMenuOpen: (open: boolean) => void;
}